CAT-1678 & CAT-1676 Implementation of two sensor for EV3

+++ catroid/src/main/java/org/catrobat/catroid/devices/mindstorms/MindstormsSensor.java
@@ -24,11 +24,11 @@ package org.catrobat.catroid.devices.mindstorms;
 
 public interface MindstormsSensor {
 
-	int getValue();
+	float getValue();
 
 	int getUpdateInterval();
 	void updateLastSensorValue();
-	int getLastSensorValue();
+	float getLastSensorValue();
 
 	String getName();
