+++ catroid/src/main/java/org/catrobat/catroid/devices/mindstorms/nxt/sensors/NXTLightSensor.java
@@ -35,7 +35,7 @@ public class NXTLightSensor extends NXTSensor {
 	}
 
 	@Override
-	public int getValue() {
+	public float getValue() {
 		return getScaledValue();
 	}
 }
